Fix 9th bug hunting problem
Casting 1 to long prevents overflowing

+++ BugHunting9/src/hunting/Main.java
@@ -9,7 +9,7 @@ public class Main {
 				int ix = valOffset + j;
 				if (rgsToRead.length == ix) break;
 				if (!rgsToRead[ix]) continue;
-				val = val | (1 << j);
+				val = val | ((long)1 << j);
 			}
 			....
 		}
